requirements.txt
streamlit
selenium
pandas
lxml
google-generativeai
streamlit-chat